refactor(core): split command_executor.py into execution_pipeline and execution_record_builder
Extracts the two concerns that pushed command_executor.py past the 800-line
ceiling (GitHub issue #810): record-building and the batch drain/retry/persist
pipeline.

- src/hassette/core/execution_record_builder.py: build_execution_record(),
  a pure function moved from CommandExecutor.build_record().
- src/hassette/core/execution_pipeline.py: RetryableBatch, enqueue_record,
  drain_and_persist, flush_queue, persist_batch, handle_fk_violation,
  emit_completion_events — module-level functions taking the executor as an
  explicit first argument, mirroring the existing resources/lifecycle.py +
  resources/operations.py delegation pattern.
- command_executor.py keeps thin delegating methods with identical
  signatures; every internal call site continues to go through
  self.<method>(...)/executor.<method>(...) bound-method lookup rather than
  a bare module-function call, since several existing tests monkeypatch
  these methods as instance attributes.

log_timeout_rate_limited stays on CommandExecutor unmoved (sketch-time
challenge finding: a single-caller ~50-line module wasn't needed to hit the
800-line target).

Pure structural move — zero test modifications, all existing tests pass
unchanged. command_executor.py: 1167 -> 799 lines.
