+++ src/hassette/core/command_executor.py
@@ -2,14 +2,12 @@
 
 import asyncio
 import contextlib
-import sqlite3
 import time
 import traceback
 import typing
 from collections.abc import Awaitable, Callable
 from contextvars import Token
-from dataclasses import dataclass, field
-from dataclasses import replace as dataclass_replace
+from dataclasses import dataclass
 from typing import ClassVar
 
 import structlog.contextvars
@@ -18,15 +16,27 @@ import uuid_utils
 from hassette.bus.error_context import BusErrorContext
 from hassette.commands import ExecuteJob, InvokeHandler
 from hassette.context import CURRENT_EXECUTION_ID
+from hassette.core import execution_pipeline
 from hassette.core.block_io_guard import MonkeypatchEvent
 from hassette.core.database_service import DatabaseService
-from hassette.core.execution_record import SYNTHETIC_ORIGIN, ExecutionRecord
+from hassette.core.execution_pipeline import (
+    # Both re-exported unused in this file — kept so tests importing from this module path
+    # (test_command_executor_pipeline_queue.py, test_command_executor_pipeline_serve.py) keep working.
+    _MAX_RETRY_COUNT as _MAX_RETRY_COUNT,
+)
+from hassette.core.execution_pipeline import (
+    _UNOWNED_WARN_RATE_LIMIT_SECS as _UNOWNED_WARN_RATE_LIMIT_SECS,
+)
+from hassette.core.execution_pipeline import (
+    RetryableBatch as RetryableBatch,
+)
+from hassette.core.execution_record import ExecutionRecord
+from hassette.core.execution_record_builder import build_execution_record
 from hassette.core.loop_watchdog import WatchdogEvent
 from hassette.core.registration import ListenerRegistration, ScheduledJobRegistration
 from hassette.core.sync_executor import SYNC_WORKER_HANDLE
 from hassette.core.telemetry.repository import TelemetryRepository
 from hassette.error_context import ErrorContext
-from hassette.events.hassette import HassetteExecutionCompletedEvent
 from hassette.exceptions import DependencyError, HassetteError
 from hassette.resources.base import Resource
 from hassette.resources.lifecycle import mark_ready
@@ -42,12 +52,8 @@ if typing.TYPE_CHECKING:
     from hassette import Hassette
     from hassette.config.classes import AppManifest
 
-_MAX_RETRY_COUNT = 3
-_UNOWNED_WARN_RATE_LIMIT_SECS = 30.0
 _TIMEOUT_WARN_SUPPRESS_SECS = 60.0
 _TIMEOUT_WARN_CACHE_MAX = 1000
-_BATCH_DRAIN_CAP = 100
-_RETRY_BACKOFF_BASE_SECONDS = 1.0
 
 
 @dataclass(frozen=True)
@@ -81,24 +87,6 @@ class ExecutionMarker:
     instance_index: int | None = None
     task_id: int | None = None
     """``id()`` of the owning ``asyncio.Task``, or ``None`` if bound outside a task."""
-
-
-@dataclass
-class RetryableBatch:
-    """A batch of records that failed to persist and should be retried.
-
-    Attributes:
-        records: Unified execution records to retry.
-        retry_count: Number of times this whole batch has been retried by the executor.
-            Unrelated to ``ExecutionRecord.retry_count`` (a per-row schema column that is
-            currently always 0); this one drives the in-memory retry/backoff loop.
-        not_before: Monotonic timestamp (time.monotonic()) before which this batch
-            must not be retried. Zero means eligible immediately.
-    """
-
-    records: list[ExecutionRecord] = field(default_factory=list)
-    retry_count: int = 0
-    not_before: float = 0.0
 
 
 class CommandExecutor(Service):
@@ -414,41 +402,8 @@ class CommandExecutor(Service):
         )
 
     def enqueue_record(self, record: ExecutionRecord) -> None:
-        """Enqueue a record, dropping and logging if the queue is full.
-
-        Also logs a WARNING when the queue exceeds the configured capacity threshold
-        (rate-limited), per lifecycle.command_executor_capacity_warn_threshold /
-        lifecycle.command_executor_capacity_warn_rate_limit_seconds.
-        """
-        max_size = self._write_queue.maxsize
-        current_size = self._write_queue.qsize()
-        lifecycle = self.hassette.config.lifecycle
-
-        # Capacity warning (rate-limited)
-        if max_size > 0 and current_size >= max_size * lifecycle.command_executor_capacity_warn_threshold:
-            now = time.monotonic()
-            if (
-                self._last_capacity_warn_ts is None
-                or now - self._last_capacity_warn_ts >= lifecycle.command_executor_capacity_warn_rate_limit_seconds
-            ):
-                self._last_capacity_warn_ts = now
-                self.logger.warning(
-                    "Write queue at %d/%d (%.0f%%) — high telemetry load",
-                    current_size,
-                    max_size,
-                    (current_size / max_size) * 100,
-                )
-
-        try:
-            self._write_queue.put_nowait(record)
-        except asyncio.QueueFull:
-            self._dropped_overflow += 1
-            self.logger.error(
-                "Write queue full (%d/%d) — dropping record (total dropped: %d)",
-                current_size,
-                max_size,
-                self._dropped_overflow,
-            )
+        """Enqueue a record, dropping and logging if the queue is full. Delegates to execution_pipeline."""
+        execution_pipeline.enqueue_record(self, record)
 
     def build_record(
         self,
@@ -457,71 +412,10 @@ class CommandExecutor(Service):
         execution_start_ts: float,
         execution_id: str,
     ) -> ExecutionRecord:
-        """Build a unified ExecutionRecord from the execution result and command.
-
-        session_id is set to None if the session hasn't been created yet (pre-Phase 1).
-        The actual session_id is injected at drain time in persist_batch.
-
-        Args:
-            cmd: The originating command.
-            result: The execution result with timing and error info.
-            execution_start_ts: Unix timestamp when execution began.
-            execution_id: UUIDv7 string for this execution instance.
-        """
-        # `result.status` is only `None` before `track_execution()` (or the CANCELLED default
-        # `_execute()` seeds before entering it) assigns a real value — every caller of
-        # `build_record()` does so after that assignment has happened. This is a runtime contract
-        # violation, not a type-exhaustiveness guard (contrast `_execute()`'s
-        # `raise AssertionError` on an unreachable `match` arm above) — a direct `build_record()`
-        # call with an unpopulated result is a real gap in the invariant, not an impossible-in-
-        # principle branch. Raise explicitly rather than silently coalescing to a fallback status,
-        # so it surfaces immediately instead of miscategorizing an execution's outcome.
-        if result.status is None:
-            raise RuntimeError("ExecutionResult.status must be populated before building a record")
-        session_id = self.hassette.try_session_id()
-
-        match cmd:
-            case InvokeHandler():
-                return ExecutionRecord(
-                    kind="handler",
-                    listener_id=cmd.listener_id,
-                    job_id=None,
-                    session_id=session_id,
-                    execution_start_ts=execution_start_ts,
-                    duration_ms=result.duration_ms,
-                    status=result.status,
-                    app_key=cmd.listener.identity.app_key,
-                    instance_index=cmd.listener.identity.instance_index,
-                    source_tier=cmd.source_tier,
-                    is_di_failure=result.is_di_failure,
-                    thread_leaked=result.thread_leaked,
-                    error_type=result.error_type,
-                    error_message=result.error_message,
-                    error_traceback=result.error_traceback,
-                    execution_id=execution_id,
-                    trigger_context_id=None if cmd.is_synthetic else cmd.event.payload.event_id,
-                    trigger_origin=SYNTHETIC_ORIGIN if cmd.is_synthetic else cmd.event.payload.origin,
-                )
-            case ExecuteJob():
-                return ExecutionRecord(
-                    kind="job",
-                    listener_id=None,
-                    job_id=cmd.job_db_id,
-                    session_id=session_id,
-                    execution_start_ts=execution_start_ts,
-                    duration_ms=result.duration_ms,
-                    status=result.status,
-                    app_key=cmd.job.app_key,
-                    instance_index=cmd.job.instance_index,
-                    source_tier=cmd.source_tier,
-                    is_di_failure=result.is_di_failure,
-                    thread_leaked=result.thread_leaked,
-                    error_type=result.error_type,
-                    error_message=result.error_message,
-                    error_traceback=result.error_traceback,
-                    execution_id=execution_id,
-                    trigger_mode=cmd.trigger_mode,
-                )
+        """Build a unified ExecutionRecord. Delegates to execution_record_builder.build_execution_record()."""
+        return build_execution_record(
+            cmd, result, execution_start_ts, execution_id, session_id=self.hassette.try_session_id()
+        )
 
     def bind_execution_context(
         self,
@@ -808,105 +702,12 @@ class CommandExecutor(Service):
         self,
         first_item: ExecutionRecord | RetryableBatch | None = None,
     ) -> None:
-        """Drain up to 100 queue items and persist them to DB.
-
-        Separates fresh ExecutionRecord items from RetryableBatch items.
-        RetryableBatch items are processed separately to preserve their retry_count.
-
-        Note: the 100-item cap applies to *queue items*, not total records.
-        A single RetryableBatch counts as 1 queue item but may contain a full
-        prior batch's worth of records.  This is acceptable for append-only
-        telemetry — a large single transaction at recovery time is benign.
-
-        Args:
-            first_item: An already-dequeued item to include as the first record.
-                When provided, at most 99 additional items are drained from the queue
-                so that the total batch size stays at 100.
-        """
-        fresh_records: list[ExecutionRecord] = []
-        retry_batches: list[RetryableBatch] = []
-
-        def _classify(item: ExecutionRecord | RetryableBatch) -> None:
-            if isinstance(item, RetryableBatch):
-                retry_batches.append(item)
-            elif isinstance(item, ExecutionRecord):
-                fresh_records.append(item)
-            else:
-                typing.assert_never(item)
-
-        if first_item is not None:
-            _classify(first_item)
-
-        # Drain remaining items up to a total batch size of _BATCH_DRAIN_CAP (non-blocking)
-        for _ in range(_BATCH_DRAIN_CAP - 1 if first_item is not None else _BATCH_DRAIN_CAP):
-            try:
-                item = self._write_queue.get_nowait()
-            except asyncio.QueueEmpty:
-                break
-            _classify(item)
-
-        # Persist fresh records as a single batch (retry_count=0)
-        if fresh_records:
-            await self.persist_batch(fresh_records)
-
-        # Process each RetryableBatch separately to preserve its retry_count.
-        # Skip batches whose backoff window has not yet elapsed — re-enqueue them.
-        now = time.monotonic()
-        for batch in retry_batches:
-            if batch.not_before > now:
-                # Backoff window still active — put it back for a later drain cycle
-                try:
-                    self._write_queue.put_nowait(batch)
-                except asyncio.QueueFull:
-                    drop_count = len(batch.records)
-                    self._dropped_overflow += drop_count
-                    self.logger.error(
-                        "Write queue full while deferring retry batch (not_before not reached) "
-                        "— dropping %d records (total overflow: %d)",
-                        drop_count,
-                        self._dropped_overflow,
-                    )
-                continue
-            await self.persist_batch(batch.records, retry_count=batch.retry_count)
+        """Drain up to 100 queue items and persist them to DB. Delegates to execution_pipeline."""
+        await execution_pipeline.drain_and_persist(self, first_item=first_item)
 
     async def flush_queue(self) -> None:
-        """Drain and persist ALL remaining items in the write queue.
-
-        Called during shutdown to ensure no records are lost.
-        Unlike drain_and_persist, there is no size limit.
-
-        Wraps persist_batch in try/except — DB may already be closed at shutdown.
-        """
-        records: list[ExecutionRecord] = []
-
-        while True:
-            try:
-                item = self._write_queue.get_nowait()
-            except asyncio.QueueEmpty:
-                break
-
-            if isinstance(item, RetryableBatch):
-                # retry_count and not_before intentionally bypassed — during shutdown,
-                # we make a single best-effort persist regardless of backoff state.
-                records.extend(item.records)
-            elif isinstance(item, ExecutionRecord):
-                records.append(item)
-            else:
-                typing.assert_never(item)
-
-        if not records:
-            return
-
-        try:
-            await self.persist_batch(records)
-        except Exception:
-            drop_count = len(records)
-            self._dropped_shutdown += drop_count
-            self.logger.error(
-                "flush_queue: failed to persist %d records during shutdown — dropped (total shutdown: %d)",
-                drop_count,
-                self._dropped_shutdown,
-            )
+        """Drain and persist ALL remaining items in the write queue. Delegates to execution_pipeline."""
+        await execution_pipeline.flush_queue(self)
 
     async def persist_batch(
         self,
@@ -914,155 +715,15 @@ class CommandExecutor(Service):
         *,
         retry_count: int = 0,
     ) -> None:
-        """Write a batch of unified execution records to the DB in a single transaction.
-
-        Session injection:
-        - Records with session_id=None are updated to the current session_id at drain time.
-        - Records with no session available are dropped with a warning.
-
-        Error classification:
-        - sqlite3.OperationalError → retry via RetryableBatch (max 3 retries).
-        - sqlite3.IntegrityError → FK violation path (row-by-row fallback).
-        - sqlite3.DataError / sqlite3.ProgrammingError → non-retryable, drop + REGRESSION log.
-        - Other Exception → non-retryable, drop + ERROR log.
-
-        Args:
-            records: Unified execution records to insert into executions.
-            retry_count: The number of times this batch has already been retried.
-        """
-        # Drain-time session_id injection
-        # Records enqueued before session creation have session_id=None.
-        # Inject the real session_id now at persist time.
-        current_session_id = self.hassette.try_session_id()
-
-        if current_session_id is not None:
-            records = [
-                dataclass_replace(r, session_id=current_session_id) if r.session_id is None else r for r in records
-            ]
-        else:
-            # Session still not ready — drop records with None session_id
-            no_session = [r for r in records if r.session_id is None]
-            if no_session:
-                self.logger.warning(
-                    "Session not yet created at drain time — dropping %d record(s) with no session_id",
-                    len(no_session),
-                )
-            records = [r for r in records if r.session_id is not None]
-
-        if not records:
-            return
-
-        try:
-            await self.hassette.database_service.submit(self.repository.persist_execution_batch(records))
-            await self.emit_completion_events(records)
-        except sqlite3.OperationalError as exc:
-            # Retryable — transient DB error (disk I/O, locked, etc.)
-            if retry_count >= _MAX_RETRY_COUNT:
-                drop_count = len(records)
-                self._dropped_exhausted += drop_count
-                self.logger.error(
-                    "Max retries (%d) exceeded for %d record(s) — dropping (total exhausted: %d): %s",
-                    _MAX_RETRY_COUNT,
-                    drop_count,
-                    self._dropped_exhausted,
-                    exc,
-                )
-            else:
-                self.logger.warning(
-                    "OperationalError persisting batch — re-enqueueing as RetryableBatch (attempt %d/%d): %s",
-                    retry_count + 1,
-                    _MAX_RETRY_COUNT,
-                    exc,
-                )
-                try:
-                    await asyncio.sleep(0)  # yield event loop before retry to avoid starving fresh records
-                    self._write_queue.put_nowait(
-                        RetryableBatch(
-                            records=list(records),
-                            retry_count=retry_count + 1,
-                            not_before=time.monotonic() + _RETRY_BACKOFF_BASE_SECONDS * (retry_count + 1),
-                        )
-                    )
-                except asyncio.QueueFull:
-                    drop_count = len(records)
-                    self._dropped_exhausted += drop_count
-                    self.logger.error(
-                        "Write queue full while re-enqueueing retry batch — dropping %d records (total exhausted: %d)",
-                        drop_count,
-                        self._dropped_exhausted,
-                    )
-
-        except sqlite3.IntegrityError:
-            # FK violation — fall back to row-by-row INSERT
-            await self.handle_fk_violation(records)
-
-        except (sqlite3.DataError, sqlite3.ProgrammingError) as exc:
-            # Non-retryable schema/data mismatch — this is a regression
-            drop_count = len(records)
-            self.logger.error(
-                "REGRESSION: Non-retryable DB error (%s) — dropping %d record(s): %s",
-                type(exc).__name__,
-                drop_count,
-                exc,
-            )
-
-        except Exception as exc:
-            # Unknown error — drop and log at ERROR
-            drop_count = len(records)
-            self.logger.error(
-                "Unexpected error persisting %d telemetry record(s) — dropping: %s",
-                drop_count,
-                exc,
-            )
+        """Write a batch of unified execution records to the DB. Delegates to execution_pipeline."""
+        await execution_pipeline.persist_batch(self, records, retry_count=retry_count)
 
     async def emit_completion_events(
         self,
         records: list[ExecutionRecord],
     ) -> None:
-        """Emit bus topic events for persisted execution records.
-
-        Fires ``HASSETTE_EVENT_EXECUTION_COMPLETED`` for each app-tier execution
-        (both handler and job kinds). The payload's ``kind`` field distinguishes
-        handler from job completions.
-
-        Payloads include ``app_key`` and ``instance_index`` sourced directly from the
-        in-memory record (populated at build time from the Listener/Job object).
-
-        Errors are suppressed so that emission failures never affect telemetry persistence.
-        """
-        try:
-            app_records = [r for r in records if r.source_tier == "app"]
-            # Regression guard: an app-tier completion should always carry an owner.
-            # An empty app_key means registration misfired (e.g. an app reload racing
-            # the meta lookup). Rate-limited since a sustained storm would otherwise
-            # log once per drain tick.
-            unowned = sum(1 for r in app_records if not r.app_key)
-            if unowned:
-                now = self._clock()
-                if (
-                    self._last_unowned_warn_ts is None
-                    or now - self._last_unowned_warn_ts >= _UNOWNED_WARN_RATE_LIMIT_SECS
-                ):
-                    self._last_unowned_warn_ts = now
-                    self.logger.warning(
-                        "Emitting %d app-tier completion event(s) with empty app_key — telemetry will be unattributed",
-                        unowned,
-                    )
-            for record in app_records:
-                exec_event = HassetteExecutionCompletedEvent.from_record(
-                    kind=record.kind,
-                    status=record.status,
-                    duration_ms=record.duration_ms,
-                    listener_id=record.listener_id,
-                    job_id=record.job_id,
-                    app_key=record.app_key,
-                    instance_index=record.instance_index,
-                    error_type=record.error_type,
-                    thread_leaked=record.thread_leaked,
-                )
-                await self.hassette.send_event(exec_event)
-        except Exception:
-            self.logger.debug("Failed to emit completion events — ignoring", exc_info=True)
+        """Emit bus topic events for persisted execution records. Delegates to execution_pipeline."""
+        await execution_pipeline.emit_completion_events(self, records)
 
     def record_blocking_event(self, event: WatchdogEvent | MonkeypatchEvent) -> None:
         """Persist a blocking event row. Must be called on the loop thread.
@@ -1134,34 +795,5 @@ class CommandExecutor(Service):
         self,
         records: list[ExecutionRecord],
     ) -> None:
-        """Handle an IntegrityError by re-inserting records with FK fallback.
-
-        Uses a single database_service.submit() call (one queue slot, one
-        transaction) to process all records row-by-row. For each record that
-        fails with an IntegrityError, the FK field is nulled and retried.
-
-        Args:
-            records: Unified execution records to insert individually.
-        """
-        try:
-            dropped = await self.hassette.database_service.submit(
-                self.repository.persist_execution_batch_with_fk_fallback(records)
-            )
-            if dropped > 0:
-                self._dropped_exhausted += dropped
-                self.logger.error(
-                    "FK violation fallback: %d record(s) dropped even with null FK (total exhausted: %d)",
-                    dropped,
-                    self._dropped_exhausted,
-                )
-            else:
-                await self.emit_completion_events(records)
-        except Exception as exc:
-            drop_count = len(records)
-            self._dropped_exhausted += drop_count
-            self.logger.error(
-                "FK violation fallback failed entirely — dropping %d record(s) (total exhausted: %d): %s",
-                drop_count,
-                self._dropped_exhausted,
-                exc,
-            )
+        """Handle an IntegrityError by re-inserting records with FK fallback. Delegates to execution_pipeline."""
+        await execution_pipeline.handle_fk_violation(self, records)
